refactor(core): split release restore planning and artifact recovery
