+++ crates/surge-core/src/releases/restore.rs
@@ -1,17 +1,17 @@
 //! Release artifact graph helpers for restore/reconstruction and pruning.
 
-use std::collections::{BTreeMap, BTreeSet};
+mod candidate;
+mod planning;
+mod recovery;
+mod retention;
+
 use std::path::Path;
 
-use crate::crypto::sha256::sha256_hex;
-use crate::error::{Result, SurgeError};
-use crate::platform::fs::write_file_atomic;
-use crate::releases::artifact_cache::{cache_path_for_key, cached_artifact_matches, fetch_or_reuse_file};
-use crate::releases::delta::{apply_delta_patch, decode_delta_patch, is_supported_delta};
-use crate::releases::manifest::{ReleaseEntry, ReleaseIndex};
-use crate::releases::version::compare_versions;
-use crate::storage::StorageBackend;
-use futures_util::stream::{self, StreamExt};
+pub use self::planning::{
+    find_previous_release_for_rid, find_release_for_version_rid, plan_full_archive_restore, sorted_releases_for_rid,
+};
+pub use self::recovery::{restore_full_archive_for_version, restore_full_archive_for_version_with_options};
+pub use self::retention::{local_checkpoint_artifacts_for_index, required_artifacts_for_index};
 
 pub type RestoreProgressCallback<'a> = dyn Fn(RestoreProgress) + Send + Sync + 'a;
 
@@ -29,597 +29,11 @@ pub struct RestoreOptions<'a> {
     pub progress: Option<&'a RestoreProgressCallback<'a>>,
 }
 
-#[derive(Debug, Clone)]
-struct ArtifactPrefetchSpec {
-    key: String,
-    sha256: String,
-    size: i64,
-}
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct RestoreArtifactSpec {
     pub key: String,
     pub sha256: String,
     pub size: i64,
-}
-
-struct RestoreCandidate<'a> {
-    base_release: &'a ReleaseEntry,
-    chain_releases: Vec<&'a ReleaseEntry>,
-    chain_deltas: Vec<crate::releases::manifest::DeltaArtifact>,
-    total_items: i64,
-    total_bytes: i64,
-    missing_items: i64,
-    missing_bytes: i64,
-}
-
-/// Return releases for a RID sorted by semantic version (oldest -> newest).
-#[must_use]
-pub fn sorted_releases_for_rid<'a>(index: &'a ReleaseIndex, rid: &str) -> Vec<&'a ReleaseEntry> {
-    let mut releases: Vec<&ReleaseEntry> = index
-        .releases
-        .iter()
-        .filter(|release| release.rid == rid || release.rid.is_empty())
-        .collect();
-    releases.sort_by(|a, b| compare_versions(&a.version, &b.version));
-    releases
-}
-
-/// Find a specific RID release by exact version.
-#[must_use]
-pub fn find_release_for_version_rid<'a>(index: &'a ReleaseIndex, rid: &str, version: &str) -> Option<&'a ReleaseEntry> {
-    sorted_releases_for_rid(index, rid)
-        .into_iter()
-        .find(|release| release.version == version)
-}
-
-/// Find the most recent release before `version` for a RID.
-#[must_use]
-pub fn find_previous_release_for_rid<'a>(
-    index: &'a ReleaseIndex,
-    rid: &str,
-    version: &str,
-) -> Option<&'a ReleaseEntry> {
-    let mut previous: Option<&ReleaseEntry> = None;
-    for release in sorted_releases_for_rid(index, rid) {
-        if compare_versions(&release.version, version) != std::cmp::Ordering::Less {
-            continue;
-        }
-        previous = Some(release);
-    }
-    previous
-}
-
-pub async fn plan_full_archive_restore(
-    storage: &dyn StorageBackend,
-    index: &ReleaseIndex,
-    rid: &str,
-    version: &str,
-) -> Result<Vec<RestoreArtifactSpec>> {
-    let releases = sorted_releases_for_rid(index, rid);
-    let target_idx = releases
-        .iter()
-        .position(|release| release.version == version)
-        .ok_or_else(|| SurgeError::NotFound(format!("Release {version} ({rid}) not found in index")))?;
-
-    for base_idx in (0..=target_idx).rev() {
-        let Some(specs) = build_restore_plan_specs(&releases, target_idx, base_idx) else {
-            continue;
-        };
-        if restore_artifacts_exist(storage, &specs).await? {
-            return Ok(specs);
-        }
-    }
-
-    Err(SurgeError::NotFound(format!(
-        "No reconstructable full archive found for {version} ({rid})"
-    )))
-}
-
-/// Restore a release full archive for `version` by RID.
-///
-/// This first attempts to download the direct full artifact. If missing, it
-/// searches for an earlier available full and rebuilds forward via delta chain.
-pub async fn restore_full_archive_for_version(
-    storage: &dyn StorageBackend,
-    index: &ReleaseIndex,
-    rid: &str,
-    version: &str,
-) -> Result<Vec<u8>> {
-    restore_full_archive_for_version_with_options(storage, index, rid, version, RestoreOptions::default()).await
-}
-
-/// Restore a release full archive with optional local-cache and progress reporting.
-///
-/// When `cache_dir` is provided, the restore path prefers any valid cached
-/// artifact chain that minimizes additional downloads rather than always
-/// fetching the direct target full archive.
-pub async fn restore_full_archive_for_version_with_options(
-    storage: &dyn StorageBackend,
-    index: &ReleaseIndex,
-    rid: &str,
-    version: &str,
-    options: RestoreOptions<'_>,
-) -> Result<Vec<u8>> {
-    let releases = sorted_releases_for_rid(index, rid);
-    let target_idx = releases
-        .iter()
-        .position(|release| release.version == version)
-        .ok_or_else(|| SurgeError::NotFound(format!("Release {version} ({rid}) not found in index")))?;
-
-    let candidate = select_restore_candidate(storage, &releases, target_idx, options.cache_dir).await?;
-
-    let report_progress = |items_done: i64, bytes_done: i64| {
-        if let Some(callback) = options.progress {
-            callback(RestoreProgress {
-                items_done,
-                items_total: candidate.total_items,
-                bytes_done,
-                bytes_total: candidate.total_bytes,
-            });
-        }
-    };
-    let prefetch_specs = {
-        let mut specs = Vec::with_capacity(candidate.chain_deltas.len().saturating_add(1));
-        specs.push(ArtifactPrefetchSpec {
-            key: candidate.base_release.full_filename.clone(),
-            sha256: candidate.base_release.full_sha256.clone(),
-            size: candidate.base_release.full_size,
-        });
-        for delta in &candidate.chain_deltas {
-            specs.push(ArtifactPrefetchSpec {
-                key: delta.filename.clone(),
-                sha256: delta.sha256.clone(),
-                size: delta.size,
-            });
-        }
-        specs
-    };
-    let used_prefetch = if let Some(cache_root) = options.cache_dir {
-        prefetch_artifacts_to_cache(
-            storage,
-            cache_root,
-            &prefetch_specs,
-            options.progress,
-            candidate.total_items,
-            candidate.total_bytes,
-        )
-        .await?;
-        true
-    } else {
-        false
-    };
-
-    let mut restored = fetch_artifact_bytes(
-        storage,
-        &candidate.base_release.full_filename,
-        &candidate.base_release.full_sha256,
-        options.cache_dir,
-    )
-    .await?;
-    verify_expected_sha256(
-        &candidate.base_release.full_sha256,
-        &restored,
-        &format!("full artifact '{}'", candidate.base_release.full_filename),
-    )?;
-
-    let mut items_done = 1i64;
-    let mut bytes_done = candidate.base_release.full_size.max(0);
-    if !used_prefetch {
-        report_progress(items_done, bytes_done);
-    }
-
-    for (release, delta) in candidate.chain_releases.iter().zip(candidate.chain_deltas.iter()) {
-        let delta_compressed = fetch_artifact_bytes(storage, &delta.filename, &delta.sha256, options.cache_dir).await?;
-        verify_expected_sha256(
-            &delta.sha256,
-            &delta_compressed,
-            &format!("delta artifact '{}'", delta.filename),
-        )?;
-        items_done = items_done.saturating_add(1);
-        bytes_done = bytes_done.saturating_add(delta.size.max(0));
-        if !used_prefetch {
-            report_progress(items_done, bytes_done);
-        }
-
-        let patch = decode_delta_patch(delta_compressed.as_slice(), delta).map_err(|_| {
-            SurgeError::NotFound(format!(
-                "Failed to decode delta artifact '{}' while restoring {version} ({rid})",
-                delta.filename
-            ))
-        })?;
-        restored = apply_delta_patch(&restored, &patch, delta).map_err(|_| {
-            SurgeError::NotFound(format!(
-                "Failed to apply delta artifact '{}' while restoring {version} ({rid})",
-                delta.filename
-            ))
-        })?;
-        verify_expected_sha256(
-            &release.full_sha256,
-            &restored,
-            &format!("rebuilt full archive for {}", release.version),
-        )?;
-        if let Some(cache_root) = options.cache_dir {
-            cache_restored_full_archive(cache_root, release, &restored)?;
-        }
-    }
-
-    Ok(restored)
-}
-
-fn cache_restored_full_archive(cache_root: &Path, release: &ReleaseEntry, restored: &[u8]) -> Result<()> {
-    let key = release.full_filename.trim();
-    if key.is_empty() {
-        return Ok(());
-    }
-
-    verify_expected_sha256(
-        &release.full_sha256,
-        restored,
-        &format!("cached rebuilt full archive for {}", release.version),
-    )?;
-    let cache_path = cache_path_for_key(cache_root, key)?;
-    write_file_atomic(&cache_path, restored)
-}
-
-async fn select_restore_candidate<'a>(
-    storage: &dyn StorageBackend,
-    releases: &[&'a ReleaseEntry],
-    target_idx: usize,
-    cache_dir: Option<&Path>,
-) -> Result<RestoreCandidate<'a>> {
-    let mut best: Option<RestoreCandidate<'a>> = None;
-    let mut cache_state = BTreeMap::new();
-    let mut storage_state = BTreeMap::new();
-
-    for base_idx in (0..=target_idx).rev() {
-        let Some(mut candidate) = build_restore_candidate(releases, target_idx, base_idx) else {
-            continue;
-        };
-        let Some((missing_items, missing_bytes)) =
-            assess_restore_candidate(storage, &candidate, cache_dir, &mut cache_state, &mut storage_state).await?
-        else {
-            continue;
-        };
-        candidate.missing_items = missing_items;
-        candidate.missing_bytes = missing_bytes;
-        if cache_dir.is_none() {
-            return Ok(candidate);
-        }
-        if best
-            .as_ref()
-            .is_none_or(|current| restore_candidate_is_better(&candidate, current))
-        {
-            best = Some(candidate);
-        }
-    }
-
-    best.ok_or_else(|| {
-        let version = releases
-            .get(target_idx)
-            .map_or_else(|| "<unknown>".to_string(), |release| release.version.clone());
-        let rid = releases
-            .get(target_idx)
-            .map_or_else(|| "<unknown>".to_string(), |release| release.rid.clone());
-        SurgeError::NotFound(format!("No reconstructable full archive found for {version} ({rid})"))
-    })
-}
-
-fn build_restore_candidate<'a>(
-    releases: &[&'a ReleaseEntry],
-    target_idx: usize,
-    base_idx: usize,
-) -> Option<RestoreCandidate<'a>> {
-    let base_release = *releases.get(base_idx)?;
-    if base_release.full_filename.trim().is_empty() {
-        return None;
-    }
-
-    let chain_releases: Vec<&ReleaseEntry> = releases
-        .iter()
-        .take(target_idx + 1)
-        .skip(base_idx + 1)
-        .copied()
-        .collect();
-    let mut chain_deltas = Vec::with_capacity(chain_releases.len());
-    let mut total_bytes = base_release.full_size.max(0);
-    for release in &chain_releases {
-        let delta = release.selected_delta()?;
-        if !is_supported_delta(&delta) {
-            return None;
-        }
-        total_bytes = total_bytes.saturating_add(delta.size.max(0));
-        chain_deltas.push(delta);
-    }
-
-    let total_items = i64::try_from(chain_deltas.len())
-        .ok()
-        .and_then(|count| count.checked_add(1))
-        .unwrap_or(i64::MAX);
-
-    Some(RestoreCandidate {
-        base_release,
-        chain_releases,
-        chain_deltas,
-        total_items,
-        total_bytes,
-        missing_items: 0,
-        missing_bytes: 0,
-    })
-}
-
-async fn assess_restore_candidate(
-    storage: &dyn StorageBackend,
-    candidate: &RestoreCandidate<'_>,
-    cache_dir: Option<&Path>,
-    cache_state: &mut BTreeMap<String, bool>,
-    storage_state: &mut BTreeMap<String, bool>,
-) -> Result<Option<(i64, i64)>> {
-    let mut missing_items = 0i64;
-    let mut missing_bytes = 0i64;
-
-    let mut specs = Vec::with_capacity(candidate.chain_deltas.len().saturating_add(1));
-    specs.push(RestoreArtifactSpec {
-        key: candidate.base_release.full_filename.clone(),
-        sha256: candidate.base_release.full_sha256.clone(),
-        size: candidate.base_release.full_size,
-    });
-    for delta in &candidate.chain_deltas {
-        specs.push(RestoreArtifactSpec {
-            key: delta.filename.clone(),
-            sha256: delta.sha256.clone(),
-            size: delta.size,
-        });
-    }
-
-    for spec in specs {
-        if let Some(cache_root) = cache_dir
-            && cached_artifact_available(cache_root, &spec, cache_state)?
-        {
-            continue;
-        }
-        if !storage_artifact_available(storage, &spec.key, storage_state).await? {
-            return Ok(None);
-        }
-        missing_items = missing_items.saturating_add(1);
-        missing_bytes = missing_bytes.saturating_add(spec.size.max(0));
-    }
-
-    Ok(Some((missing_items, missing_bytes)))
-}
-
-fn cached_artifact_available(
-    cache_root: &Path,
-    spec: &RestoreArtifactSpec,
-    cache_state: &mut BTreeMap<String, bool>,
-) -> Result<bool> {
-    if let Some(cached) = cache_state.get(&spec.key) {
-        return Ok(*cached);
-    }
-
-    let cache_path = cache_path_for_key(cache_root, &spec.key)?;
-    let cached = cached_artifact_matches(&cache_path, &spec.sha256)?;
-    cache_state.insert(spec.key.clone(), cached);
-    Ok(cached)
-}
-
-async fn storage_artifact_available(
-    storage: &dyn StorageBackend,
-    key: &str,
-    storage_state: &mut BTreeMap<String, bool>,
-) -> Result<bool> {
-    if let Some(cached) = storage_state.get(key) {
-        return Ok(*cached);
-    }
-
-    let available = match storage.head_object(key).await {
-        Ok(_) => true,
-        Err(SurgeError::NotFound(_)) => false,
-        Err(e) => return Err(e),
-    };
-    storage_state.insert(key.to_string(), available);
-    Ok(available)
-}
-
-fn restore_candidate_is_better(candidate: &RestoreCandidate<'_>, current: &RestoreCandidate<'_>) -> bool {
-    (
-        candidate.missing_bytes,
-        candidate.missing_items,
-        candidate.chain_deltas.len(),
-    ) < (current.missing_bytes, current.missing_items, current.chain_deltas.len())
-}
-
-async fn fetch_artifact_bytes(
-    storage: &dyn StorageBackend,
-    key: &str,
-    expected_sha256: &str,
-    cache_dir: Option<&Path>,
-) -> Result<Vec<u8>> {
-    if let Some(cache_root) = cache_dir {
-        let cache_path = cache_path_for_key(cache_root, key)?;
-        fetch_or_reuse_file(storage, key, &cache_path, expected_sha256, None).await?;
-        return std::fs::read(cache_path).map_err(SurgeError::Io);
-    }
-    storage.get_object(key).await
-}
-
-async fn prefetch_artifacts_to_cache(
-    storage: &dyn StorageBackend,
-    cache_root: &Path,
-    specs: &[ArtifactPrefetchSpec],
-    progress: Option<&RestoreProgressCallback<'_>>,
-    items_total: i64,
-    bytes_total: i64,
-) -> Result<()> {
-    const PREFETCH_CONCURRENCY: usize = 4;
-
-    let mut items_done = 0i64;
-    let mut bytes_done = 0i64;
-    let mut prefetch_stream = stream::iter(specs.iter().cloned())
-        .map(|spec| async move {
-            let cache_path = cache_path_for_key(cache_root, &spec.key)?;
-            fetch_or_reuse_file(storage, &spec.key, &cache_path, &spec.sha256, None).await?;
-            Ok::<i64, SurgeError>(spec.size.max(0))
-        })
-        .buffer_unordered(PREFETCH_CONCURRENCY);
-
-    while let Some(result) = prefetch_stream.next().await {
-        let size = result?;
-        items_done = items_done.saturating_add(1);
-        bytes_done = bytes_done.saturating_add(size);
-        if let Some(callback) = progress {
-            callback(RestoreProgress {
-                items_done,
-                items_total,
-                bytes_done,
-                bytes_total,
-            });
-        }
-    }
-
-    Ok(())
-}
-
-/// Compute the required artifact keys for a release index after dependency
-/// pruning. This keeps the minimum forward chain for each RID from one base
-/// full artifact plus deltas.
-#[must_use]
-pub fn required_artifacts_for_index(index: &ReleaseIndex) -> BTreeSet<String> {
-    let mut by_rid: BTreeMap<&str, Vec<&ReleaseEntry>> = BTreeMap::new();
-    for release in &index.releases {
-        by_rid.entry(release.rid.as_str()).or_default().push(release);
-    }
-
-    let mut required = BTreeSet::new();
-    for releases in by_rid.values_mut() {
-        releases.sort_by(|a, b| compare_versions(&a.version, &b.version));
-        extend_required_artifacts_for_sorted_releases(releases, &mut required);
-    }
-    required
-}
-
-#[must_use]
-pub fn local_checkpoint_artifacts_for_index(index: &ReleaseIndex, per_rid_limit: usize) -> BTreeSet<String> {
-    let mut by_rid: BTreeMap<&str, Vec<&ReleaseEntry>> = BTreeMap::new();
-    for release in &index.releases {
-        by_rid.entry(release.rid.as_str()).or_default().push(release);
-    }
-
-    let mut retained = BTreeSet::new();
-    for releases in by_rid.values_mut() {
-        releases.sort_by(|a, b| compare_versions(&a.version, &b.version));
-        for release in releases.iter().rev().take(per_rid_limit) {
-            let key = release.full_filename.trim();
-            if !key.is_empty() {
-                retained.insert(key.to_string());
-            }
-        }
-    }
-
-    retained
-}
-
-fn extend_required_artifacts_for_sorted_releases(releases: &[&ReleaseEntry], required: &mut BTreeSet<String>) {
-    if releases.is_empty() {
-        return;
-    }
-
-    let mut required_full_indices = Vec::new();
-    if let Some(first_full_idx) = releases
-        .iter()
-        .position(|release| !release.full_filename.trim().is_empty())
-    {
-        required_full_indices.push(first_full_idx);
-    }
-
-    for (idx, release) in releases.iter().enumerate().skip(1) {
-        if release.selected_delta().is_none() && !release.full_filename.trim().is_empty() {
-            required_full_indices.push(idx);
-        }
-    }
-
-    if required_full_indices.is_empty() {
-        return;
-    }
-
-    required_full_indices.sort_unstable();
-    required_full_indices.dedup();
-
-    for idx in &required_full_indices {
-        let full = releases[*idx].full_filename.trim();
-        if !full.is_empty() {
-            required.insert(full.to_string());
-        }
-    }
-
-    let first_required_full = required_full_indices[0];
-    for release in releases.iter().skip(first_required_full + 1) {
-        for delta in release.all_deltas() {
-            let key = delta.filename.trim();
-            if !key.is_empty() {
-                required.insert(key.to_string());
-            }
-        }
-    }
-}
-
-fn verify_expected_sha256(expected: &str, data: &[u8], context: &str) -> Result<()> {
-    let expected = expected.trim();
-    if expected.is_empty() {
-        return Ok(());
-    }
-
-    let actual = sha256_hex(data);
-    if actual != expected {
-        return Err(SurgeError::Storage(format!(
-            "SHA-256 mismatch for {context}: expected {expected}, got {actual}"
-        )));
-    }
-    Ok(())
-}
-
-fn build_restore_plan_specs(
-    releases: &[&ReleaseEntry],
-    target_idx: usize,
-    base_idx: usize,
-) -> Option<Vec<RestoreArtifactSpec>> {
-    let base_release = releases.get(base_idx)?;
-    let full_key = base_release.full_filename.trim();
-    if full_key.is_empty() {
-        return None;
-    }
-
-    let mut specs = vec![RestoreArtifactSpec {
-        key: full_key.to_string(),
-        sha256: base_release.full_sha256.clone(),
-        size: base_release.full_size,
-    }];
-
-    for release in releases.iter().take(target_idx + 1).skip(base_idx + 1) {
-        let delta = release.selected_delta()?;
-        if !is_supported_delta(&delta) {
-            return None;
-        }
-        specs.push(RestoreArtifactSpec {
-            key: delta.filename.clone(),
-            sha256: delta.sha256.clone(),
-            size: delta.size,
-        });
-    }
-
-    Some(specs)
-}
-
-async fn restore_artifacts_exist(storage: &dyn StorageBackend, specs: &[RestoreArtifactSpec]) -> Result<bool> {
-    for spec in specs {
-        match storage.head_object(&spec.key).await {
-            Ok(_) => {}
-            Err(SurgeError::NotFound(_)) => return Ok(false),
-            Err(e) => return Err(e),
-        }
-    }
-
-    Ok(true)
 }
 
 #[cfg(test)]
@@ -628,9 +42,12 @@ mod tests {
 
     use super::*;
     use crate::archive::packer::ArchivePacker;
+    use crate::crypto::sha256::sha256_hex;
     use crate::diff::chunked::{ChunkedDiffOptions, chunked_bsdiff};
     use crate::diff::wrapper::bsdiff_buffers;
-    use crate::releases::manifest::{DeltaArtifact, ReleaseEntry};
+    use crate::releases::artifact_cache::cache_path_for_key;
+    use crate::releases::manifest::{DeltaArtifact, ReleaseEntry, ReleaseIndex};
+    use crate::storage::StorageBackend;
     use crate::storage::filesystem::FilesystemBackend;
 
     fn make_entry(version: &str) -> ReleaseEntry {
